test: migrate agent-run-telemetry to bdd + clean up legacy composes
